+++ src/UI/Controllers/UserGroupController.cs
@@ -6,7 +6,10 @@ using CodeCampServer.Core.Domain.Model;
 using CodeCampServer.UI.Helpers.Filters;
 using CodeCampServer.UI.Helpers.Mappers;
 using CodeCampServer.UI.Models.Forms;
+using CodeCampServer.Core.Services.Impl;
 using MvcContrib;
+using CodeCampServer.UI;
+using CodeCampServer.Core.Services;
 
 
 namespace CodeCampServer.UI.Controllers
@@ -15,15 +18,13 @@ namespace CodeCampServer.UI.Controllers
 	{
 		private readonly IUserGroupRepository _repository;
 		private readonly IUserGroupMapper _mapper;
-	    private readonly IConferenceRepository _conferenceRepository;
-	    private readonly IConferenceMapper _conferenceMapper;
+        private readonly ISecurityContext _securityContext;
 
-	    public UserGroupController(IUserGroupRepository repository, IUserGroupMapper mapper,IConferenceRepository conferenceRepository,IConferenceMapper conferenceMapper) : base(repository, mapper)
+	    public UserGroupController(IUserGroupRepository repository, IUserGroupMapper mapper,IConferenceRepository conferenceRepository,IConferenceMapper conferenceMapper, ISecurityContext securityContext) : base(repository, mapper)
 		{
 			_repository = repository;
 			_mapper = mapper;
-		    _conferenceRepository = conferenceRepository;
-	        _conferenceMapper = conferenceMapper;
+	        _securityContext = securityContext;
 		}
 
 		
@@ -46,38 +47,57 @@ namespace CodeCampServer.UI.Controllers
 			return View(entityListDto);
 		}
 		
-        [RequireAdminAuthorizationFilter]
+        //[RequireAdminAuthorizationFilter]
 		public ActionResult Edit(Guid Id)
 		{
+
 			if (Id == Guid.Empty)
 			{
 				TempData.Add("message", "UserGroup has been deleted.");
 				return RedirectToAction<UserGroupController>(c => c.List());
 			}
+
             UserGroup userGroup = _repository.GetById(Id);
-			return View(_mapper.Map(userGroup));
+            if(!CurrentUserHasPermissionToEditUserGroup(userGroup))
+            {
+                return View(ViewPages.NotAuthorized);
+            }
+            return View(_mapper.Map(userGroup));
 		}
 
-        
-        [RequireAdminAuthorizationFilter]
-        [ValidateInput(false)] 
+	    protected bool CurrentUserHasPermissionToEditUserGroup(Guid Id)
+	    {
+	        return CurrentUserHasPermissionToEditUserGroup(_repository.GetById(Id));
+	    }
+
+        protected bool CurrentUserHasPermissionToEditUserGroup(UserGroup userGroup)
+        {
+            return _securityContext.HasPermissionsFor(userGroup);
+        }
+
+	    [ValidateInput(false)] 
 		[ValidateModel(typeof (UserGroupForm))]
 		public ActionResult Save([Bind(Prefix = "")] UserGroupForm form)
 		{
-			return ProcessSave(form, entity => RedirectToAction<UserGroupController>(c => c.List()));
+            if(_securityContext.HasPermissionsForUserGroup(form.Id))
+            {
+                return ProcessSave(form, entity => RedirectToAction<UserGroupController>(c => c.List()));
+            }
+	        return View(ViewPages.NotAuthorized);
 		}
+        
 
 		protected override IDictionary<string, string[]> GetFormValidationErrors(UserGroupForm form)
 		{
 			var result = new ValidationResult();
-			if (UserGroupKeyAlreadyExists(form))
+			if (CurrentUserHasPermissionToEditUserGroup(form.Id) && UserGroupKeyAlreadyExists(form))
 			{
 				result.AddError<UserGroupForm>(x => x.Key, "This entity key already exists");
 			}
 			return result.GetAllErrors();
 		}
 
-		private bool UserGroupKeyAlreadyExists(UserGroupForm message)
+	    private bool UserGroupKeyAlreadyExists(UserGroupForm message)
 		{
 			UserGroup entity = _repository.GetByKey(message.Key);
 			return entity != null && entity.Id != message.Id;
@@ -93,6 +113,11 @@ namespace CodeCampServer.UI.Controllers
         [RequireAdminAuthorizationFilter]
         public ActionResult Delete(UserGroup entity)
         {
+            if (!CurrentUserHasPermissionToEditUserGroup(entity))
+            {
+                return View(ViewPages.NotAuthorized);
+            }
+
             if (entity.GetUsers().Length == 0)
             {
                 _repository.Delete(entity);
